requirements.txt
python-firebase==1.2
Flask==0.12flask-cors